Flight classification
This script determines for every data point whether the stork was flying and whether it was climbing or gliding.
Python script, this one got majorly updated after I translated it from R. 
I'm not sure how easy it is to translate this function back into R-code if you don't have Python experience. I will make a new file here and translate little bits at a time, but it may take a while.